move template to reatures
